pyproject.toml
[project]
name = "tastie"
version = "0.1.0"
description = "a personal bookmark repository with optional sharing"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi==0.128.0",
    "uvicorn[standard]==0.40.0",
    "jinja2==3.1.6",
    "sqlalchemy[asyncio]==2.0.45",
    "aiosqlite==0.22.1",
    "python-multipart==0.0.21",
    "argon2-cffi==25.1.0",
    "itsdangerous==2.2.0",
    "slowapi==0.1.9",
    "email-validator==2.3.0",
    "python-dotenv==1.2.1",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pip-audit>=2.9.0",
    "rcssmin>=1.1.0",
    "rjsmin>=1.2.0",
]
lint = [
    "ruff>=0.14.10",
]
